The Javadoc for properties such as NODE_ALIAS_M has been enhanced, and a RelationshipExtractor has been added to include the relevant methods
